+++ src/main/kotlin/days/Day5.kt
@@ -5,7 +5,7 @@ import common.stackOf
 
 class Day5 : Day(5) {
 
-    private fun program() = inputString.split(",").map { it.trim().toInt() }.toMutableList()
+    private fun program() = inputString.split(",").map { it.trim().toLong() }.toMutableList()
 
     override fun partOne(): Any {
         return IntcodeComputer(program().toMutableList()).runWithIO(stackOf(1)).last()
